Servicios de gestión de subida de imágenes terminados, sube ficheros perfectamente

+++ APP_BACKEND/controllers/file.controller.js
@@ -10,53 +10,59 @@ const getFiles = async (req, res, next) => {
     const projectId = req.params['id'];
     const files = await fileService.getAllFromProject(projectId);
 
-    if (files.length === 0)
+    const responses = await Promise.all(files.map(async (file) => {
+        let img = await fileUtils.getFile(file.nombre);
+        
+        return {
+            ...file,
+            imgData: img
+        };
+    }));
+
+    if (responses.length === 0)
         return responseMessage(res, 200, i18n.__('files.noAttachedFiles'));
 
-    return responseMessage(res, 200, files);
+    return responseMessage(res, 200, responses);
 };
 
 const postFile = async (req, res, next) => {
     const projectId = req.params['id'];
-    const { descripcion } = req.body;
-    const file = req.files.file;
-    console.log(file);
-
-    const data = fileUtils.uploadFile(file);
-    console.log(data);
-    if (!data.filename && !data.filepath)
+    const { data, fileName, descripcion, alternativo } = req.body;
+    
+    const resp = fileUtils.uploadFile({ data, filename: fileName });
+    if (!resp.filename && !resp.filepath)
         return responseError(res, 400, i18n.__('files.missingInfo'));
 
-    const { filename, filepath } = data;
-    //console.log(data);
+    const { filename, filepath } = resp;
 
     const response = await fileService.postFileToProject({
         projectId,
         description: descripcion,
         filename,
         filepath,
+        alternativo
     });
-    console.log(response);
 
-    if (response.affectedRows <= 0)
+    if (response.affectedRows === 0)
         return responseError(res, 400, i18n.__('files.uploadError'));
-
-    const fileId = response.insertId;
 
     return responseMessage(res, 200, i18n.__('files.uploadSuccess'));
 };
 
 const putFile = async (req, res, next) => {
     const { projectId, fileId } = req.params;
-    const { descripcion } = req.body;
-    const file = req.files.file;
+    const { data, fileName, descripcion, alternativo } = req.body;
 
-    const data = fileUtils.uploadFile(file);
-    console.log(data);
-    if (!data.filename && !data.filepath)
+    const currentFile = await fileService.getOne(fileId);
+
+    if (!currentFile)
+        return responseError(res, 400, i18n.__('files.notFound'));
+
+    const resp = fileUtils.uploadFile({ data, filename: fileName });
+    if (!resp.filename && !resp.filepath)
         return responseError(res, 400, i18n.__('files.missingInfo'));
 
-    const { filename, filepath } = data;
+    const { filename, filepath } = resp;
 
     const response = await fileService.updateFileOnProject({
         projectId, 
@@ -64,21 +70,30 @@ const putFile = async (req, res, next) => {
         description: descripcion,
         filename,
         filepath,
+        alternativo,
     });
-    console.log(response);
 
     if (response.affectedRows <= 0)
         return responseError(res, 400, i18n.__('files.updateError'));
+
+    fileUtils.removeFileFromDirectory(currentFile.nombre);
 
     return responseMessage(res, 200, i18n.__('files.updateSuccess'));
 };
 
 const deleteFile = async (req, res, next) => {
     const { projectId, fileId } = req.params;
+    const file = await fileService.getOne(fileId);
+
+    if (!file)
+        return responseError(res, 400, i18n.__('files.notFound'));
+
     const response = await fileService.deleteOneFromProject({ projectId, fileId });
     
     if (response.affectedRows <= 0)
         return responseError(res, 400, i18n.__('files.removeError'));
+
+    fileUtils.removeFileFromDirectory(file.nombre);
 
     return responseMessage(res, 200, i18n.__('files.removeSuccess'));
 };
